# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cms/posts:
    get:
      operationId: listCmsPosts
      tags:
        - CMS Posts
      summary: List posts (admin)
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PostStatus'
        - in: query
          name: authorId
          schema:
            type: string
            format: uuid
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Page of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCmsPost
      tags:
        - CMS Posts
      summary: Create a draft post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cms/posts/{id}:
    get:
      operationId: getCmsPost
      tags:
        - CMS Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateCmsPost
      tags:
        - CMS Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostBody'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteCmsPost
      tags:
        - CMS Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted (soft delete)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /cms/posts/{id}/publish:
    post:
      operationId: publishCmsPost
      tags:
        - CMS Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /cms/posts/{id}/schedule:
    post:
      operationId: scheduleCmsPost
      tags:
        - CMS Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scheduledFor
              properties:
                scheduledFor:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /cms/posts/{id}/archive:
    post:
      operationId: archiveCmsPost
      tags:
        - CMS Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details: {}
...
    PostStatus:
      type: string
      enum:
        - draft
        - scheduled
        - published
        - archived
...
    Category:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
...
    Tag:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
...
    AuthorSummary:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
    Post:
      type: object
      required:
        - id
        - slug
        - title
        - status
        - authorId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        bodyHtml:
          type: string
          nullable: true
        bodyMarkdown:
          type: string
          nullable: true
        excerpt:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        ogImageUrl:
          type: string
          nullable: true
        authorId:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/AuthorSummary'
        status:
          $ref: '#/components/schemas/PostStatus'
        publishedAt:
          type: string
          format: date-time
          nullable: true
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        seoCanonicalUrl:
          type: string
          nullable: true
        readingTimeMin:
          type: integer
          nullable: true
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    CreatePostBody:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        bodyHtml:
          type: string
          nullable: true
        bodyMarkdown:
          type: string
          nullable: true
        excerpt:
          type: string
          nullable: true
        heroImageId:
          type: string
          format: uuid
          nullable: true
        ogImageId:
          type: string
          format: uuid
          nullable: true
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        seoCanonicalUrl:
          type: string
          nullable: true
        readingTimeMin:
          type: integer
          nullable: true
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
        tagIds:
          type: array
          items:
            type: string
            format: uuid
    UpdatePostBody:
      type: object
      properties:
        title:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        bodyHtml:
          type: string
          nullable: true
        bodyMarkdown:
          type: string
          nullable: true
        excerpt:
          type: string
          nullable: true
        heroImageId:
          type: string
          format: uuid
          nullable: true
        ogImageId:
          type: string
          format: uuid
          nullable: true
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        seoCanonicalUrl:
          type: string
          nullable: true
        readingTimeMin:
          type: integer
          nullable: true
        categoryIds:
          type: array
          nullable: true
          items:
            type: string
            format: uuid
        tagIds:
          type: array
          nullable: true
          items:
            type: string
            format: uuid
...